将 itinerary 改为int

+++ src/common/db.h
@@ -10,7 +10,7 @@
 
 #include "object.h"
 #include "macro.h"
- 
+
 #ifdef __cplusplus
     extern "C"{
 #endif
@@ -23,7 +23,7 @@
 #define DB_NAME "gps"
 #define MAX_QUERY 400
 
-typedef int (*ONEITINERARY_PROC)(int starttime, double startlat, double startlon, int endtime, double endlat, double endlon, short itinerary, void *userdata);
+typedef int (*ONEITINERARY_PROC)(int starttime, double startlat, double startlon, int endtime, double endlat, double endlon, int itinerary, void *userdata);
 typedef int (*ONEGPS_PROC)(int timestamp, double latitude, double longitude, char speed, short course, void *userdata);
 
 int db_initial(void);
@@ -37,7 +37,7 @@ int db_saveGPS(const char* imeiName, int timestamp, float lat, float lon, char s
 int db_getGPS(const char *imeiName, int starttime, int endtime, void *action, void *userdata);
 int db_getItinerary(const char *imeiName, int starttime, int endtime, void *action, void *userdata);
 int db_createItinerary(const char* tableName);
-int db_saveItinerary(const char* tableName, int starttime, float startlat, float startlon, int endtime, float endlat, float endlon, short itinerary);
+int db_saveItinerary(const char* tableName, int starttime, float startlat, float startlon, int endtime, float endlat, float endlon, int itinerary);
 int db_saveCGI(const char* imeiName, int timestamp, const CGI_MC cell[], int cellNo);
 int db_deleteTelNumber(const char *imeiName);
 int db_replaceTelNumber(const char *imeiName, const char *telNumber);
